refactor: add descriptionComponents
