+++ reactjs-films-homework/src/pages/ActorProfile/ActorDescription/styles.ts
@@ -1,0 +1,8 @@
+import { makeStyles } from "@mui/styles";
+
+
+export const useStyles = makeStyles(() => ({
+  descriptionItem: {
+    marginBottom: '20px'
+  },
+}))
