+++ reactjs-films-homework/src/pages/MovieDetails/MovieDescription/styles.ts
@@ -1,0 +1,7 @@
+import { makeStyles } from "@mui/styles";
+
+export const useStyles = makeStyles(() => ({
+  descriptionItem: {
+    marginBottom: '20px'
+  },
+}))
